refactor: 症状ワークシート編集画面の切り出し
